ml_scripts/requirements_ml.txt
pandas
scikit-learn
snowflake-connector-python[pandas]
joblib